simple-nebula/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
drf-nested-routers==0.93.5
psycopg2-binary==2.9.10
python-dotenv==1.1.0
gunicorn==23.0.0
dj-database-url==2.1.0
django-redis==5.4.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
cryptography==42.0.5
pyOpenSSL==24.0.0
PyYAML==6.0.2
jinja2==3.1.3 